+++ Assets/Scripts/PlaneVerbStructs.cs
@@ -14,26 +14,26 @@ namespace GPUVerb
     [StructLayout(LayoutKind.Sequential)]
     public struct PlaneVerbVec2
     {
-        public float x;
-        public float y;
+        public double x;
+        public double y;
 
-        public PlaneVerbVec2(float x, float y) 
+        public PlaneVerbVec2(double x, double y) 
         {
             this.x = x;
             this.y = y;
         }
 
-        public static implicit operator Vector2(PlaneVerbVec2 vec2) => new Vector2(vec2.x, vec2.y);
+        public static implicit operator Vector2(PlaneVerbVec2 vec2) => new Vector2((float)(vec2.x), (float)(vec2.y));
         public static implicit operator PlaneVerbVec2(Vector2 vec2) => new PlaneVerbVec2(vec2.x, vec2.y);
     }
 
     [StructLayout(LayoutKind.Sequential)]
     public struct PlaneVerbOutput
     {
-        public float occlusion;
-        public float wetGain;
-        public float rt60;
-        public float lowpass;
+        public double occlusion;
+        public double wetGain;
+        public double rt60;
+        public double lowpass;
         public PlaneVerbVec2 direction;
         public PlaneVerbVec2 sourceDirectivity;
     }
@@ -42,26 +42,26 @@ namespace GPUVerb
     public struct PlaneVerbAABB
     {
         public PlaneVerbVec2 position;
-        public float width;
-        public float height;
-        public float absorption;
+        public double width;
+        public double height;
+        public double absorption;
         public Vector2 min
         {
-            get => new Vector2(position.x - width / 2, position.y - height / 2);
+            get => new Vector2((float)(position.x - width / 2), (float)(position.y - height / 2));
         }
         public Vector2 max
         {
-            get => new Vector2(position.x + width / 2, position.y + height / 2);
+            get => new Vector2((float)(position.x + width / 2), (float)(position.y + height / 2));
         }
 
-        public PlaneVerbAABB(PlaneVerbVec2 position, float width, float height, float absorption)
+        public PlaneVerbAABB(PlaneVerbVec2 position, double width, double height, double absorption)
         {
             this.position = position;
             this.width = width;
             this.height = height;
             this.absorption = absorption;
         }
-        public PlaneVerbAABB(Bounds bounds, float absorption)
+        public PlaneVerbAABB(Bounds bounds, double absorption)
         {
             this.position = new PlaneVerbVec2(bounds.center.x, bounds.center.z);
             this.width = bounds.size.x;
